Changed buttons in navbar to navlinks. Changed fmainnav file to css module
